Patient

+++ PharmacyProjectBack/src/main/java/com/example/demo/repository/UserRepository/PatientRepository.java
@@ -7,6 +7,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 
 
-public interface PatientRepository  extends JpaRepository<Patient,Integer>{
-    
+public interface PatientRepository  extends JpaRepository<Patient,Long>{
 }
